requirements.txt
async-lru==1.0.2
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
idna==2.10
iniconfig==1.1.1
loguru==0.5.3
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyaes==1.6.1
pyparsing==2.4.7
Pyrogram==1.2.6
PySocks==1.7.1
pytest==6.2.3
python-dotenv==0.16.0
requests==2.25.1
TgCrypto==1.2.2
toml==0.10.2
urllib3==1.26.4
win32-setctime==1.0.3
